backend/ml/requirements.txt
fastapi>=0.100.0
python-dotenv>=1.0.0
uvicorn>=0.22.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
joblib>=1.3.0
mysql-connector-python>=8.0.0
deep-translator>=1.11.0
